requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
python-docx>=0.8.11
streamlit>=1.30.0
pytesseract>=0.3.10
